requirements_free.txt
python-dotenv==1.0.1 # For reading environment variables stored in .env file
langchain==0.2.2
langchain-community==0.2.3
unstructured==0.14.4 # Document loading
# onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
chromadb==0.5.0 # Vector storage
tiktoken==0.7.0  # For embeddings 

# Free alternatives to OpenAI
sentence-transformers==5.0.0 # For free embeddings
huggingface-hub==0.34.1 # For Hugging Face API access
requests==2.31.0 # For API calls

# install markdown dependencies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out. 